# mydoc/export/agent-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: VCP Agent Gateway API
  version: 1.0.0
  description: |-
    VCP Agent Gateway 的正式 Native published contract。

    说明：
    - 当前对外协议版本为 `v1`，发布阶段为 `ga`。
    - 该文档覆盖当前正式开放的 `health / capabilities / agents / metrics / memory / context / tools / jobs / events` 资源。
    - 推荐优先使用独立 Gateway 凭证接入；Basic Auth 与 `admin_auth` Cookie 仅作为过渡兼容链路记录。
    - 该契约与 `routes/agentGatewayRoutes.js`、M7-M11 主 specs 以及 contract validation 保持对齐。
  contact:
    name: VCP Team
  x-gateway-version: v1
  x-release-stage: ga
  x-vcp-version: 1.0.0
servers:
  - url: http://localhost:3000
    description: 本地默认端口
  - url: http://localhost:8080
    description: 备用端口
...
security:
  - gatewayKeyHeader: []
  - gatewayBearerAuth: []
  - basicAuth: []
  - adminAuthCookie: []
paths:
...
  /agent_gateway/tools/{toolName}/invoke:
    post:
      tags:
        - Tools
      summary: 调用单个 tool
      operationId: invokeAgentGatewayTool
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInvokeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ToolInvokeSuccess"
        "202":
          $ref: "#/components/responses/ToolInvokeAccepted"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/jobs/{jobId}:
    get:
      tags:
        - Jobs
      summary: 轮询正式 job runtime 状态
      operationId: getAgentGatewayJob
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/AgentIdQueryOptional"
        - $ref: "#/components/parameters/SessionIdQuery"
        - $ref: "#/components/parameters/RequestIdQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/RuntimeQuery"
      responses:
        "200":
          $ref: "#/components/responses/JobSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/jobs/{jobId}/cancel:
    post:
      tags:
        - Jobs
      summary: 取消一个可取消的 job
      operationId: cancelAgentGatewayJob
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCancelRequest"
      responses:
        "200":
          $ref: "#/components/responses/JobSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/events/stream:
    get:
      tags:
        - Events
      summary: 通过 SSE 读取 runtime 事件流
      operationId: streamAgentGatewayEvents
      parameters:
        - $ref: "#/components/parameters/JobIdQuery"
        - $ref: "#/components/parameters/AgentIdQueryOptional"
        - $ref: "#/components/parameters/SessionIdQuery"
        - $ref: "#/components/parameters/RequestIdQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/RuntimeQuery"
      responses:
        "200":
          description: SSE 事件流
          headers:
            x-agent-gateway-trace-id:
              $ref: "#/components/headers/XTraceId"
            x-agent-gateway-version:
              $ref: "#/components/headers/XGatewayVersion"
          content:
            text/event-stream:
              schema:
                type: string
                example: >
                  event: gateway.meta

                  data:
                  {"requestId":"req-stream-001","gatewayVersion":"v1","authMode":"gateway_key","authSource":"x-agent-gateway-key","gatewayId":"gw-prod"}


                  event: job.waiting_approval

                  data:
                  {"eventId":"evt_001","eventType":"job.waiting_approval","jobId":"job_001","requestId":"req-stream-001","agentId":"Ariadne","sessionId":"sess-001","gatewayId":"gw-prod","timestamp":"2026-04-19T12:00:00.000Z","data":{"status":"waiting_approval","operation":"tool.invoke","metadata":{"toolName":"ProtectedTool"}}}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    gatewayKeyHeader:
      type: apiKey
      in: header
      name: x-agent-gateway-key
      description: 推荐的 Native Gateway 专用凭证。
    gatewayBearerAuth:
      type: http
      scheme: bearer
      description: 与 gateway key 等价的 Bearer 认证输入。
    basicAuth:
      type: http
      scheme: basic
      description: 过渡期兼容的管理后台 Basic Auth。
    adminAuthCookie:
      type: apiKey
      in: cookie
      name: admin_auth
      description: 过渡期兼容的管理态 Cookie。
  headers:
    XRequestId:
      description: 请求 ID
      schema:
        type: string
    XGatewayVersion:
      description: Agent Gateway 运行时版本
      schema:
        type: string
        example: v1
    XTraceId:
      description: Agent Gateway 操作级 trace 标识
      schema:
        type: string
        example: agwop_1234567890abcdef123456
  parameters:
...
    JobIdPath:
      name: jobId
      in: path
      required: true
      schema:
        type: string
    ToolNamePath:
      name: toolName
      in: path
      required: true
      schema:
        type: string
...
    AgentIdQueryOptional:
      name: agentId
      in: query
      required: false
      schema:
        type: string
    JobIdQuery:
      name: jobId
      in: query
      required: false
      schema:
        type: string
    RequestIdQuery:
      name: requestId
      in: query
      required: false
      schema:
        type: string
    SessionIdQuery:
      name: sessionId
      in: query
      required: false
      schema:
        type: string
    SourceQuery:
      name: source
      in: query
      required: false
      schema:
        type: string
    RuntimeQuery:
      name: runtime
      in: query
      required: false
      schema:
        type: string
  responses:
...
    ToolInvokeSuccess:
      description: 工具已同步完成
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ToolInvokeEnvelope"
    ToolInvokeAccepted:
      description: 工具调用已进入 deferred runtime
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ToolInvokeEnvelope"
    JobSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JobEnvelope"
...
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TooManyRequests:
      description: 请求超过当前网关运营保护阈值
      headers:
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            rateLimited:
              summary: rate limit rejection
              value:
                success: false
                error: Request rate limit exceeded for this operation
                code: AGW_RATE_LIMITED
                details:
                  operationName: metrics.read
                  traceId: agwop_429example1234567890abcd
                  reason: rate_limited
                  retryAfterMs: 3000
                  limit: 5
                  windowMs: 60000
                meta:
                  requestId: req-metrics-rate-001
                  durationMs: 1
                  gatewayVersion: v1
                  traceId: agwop_429example1234567890abcd
                  operationName: metrics.read
                  retryAfterMs: 3000
    PayloadTooLarge:
      description: 请求体超过当前 operation 的 payload 上限
      headers:
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            payloadRejected:
              summary: payload size rejection
              value:
                success: false
                error: Request payload exceeds the configured operation limit
                code: AGW_PAYLOAD_TOO_LARGE
                details:
                  operationName: tool.invoke
                  traceId: agwop_413example1234567890abcd
                  reason: payload_too_large
                  payloadBytes: 16384
                  maxPayloadBytes: 4096
                meta:
                  requestId: req-tool-payload-001
                  durationMs: 0
                  gatewayVersion: v1
                  traceId: agwop_413example1234567890abcd
                  operationName: tool.invoke
                  retryAfterMs: 0
    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: 被共享 policy 或当前权限限制拒绝
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 目标 agent、tool、job 或资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: 请求与当前资源状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    GatewayMeta:
      type: object
      required:
        - requestId
        - durationMs
        - gatewayVersion
      properties:
        requestId:
          type: string
        durationMs:
          type: integer
        gatewayVersion:
          type: string
          example: v1
        authMode:
          type: string
          example: gateway_key
        authSource:
          type: string
          example: x-agent-gateway-key
        gatewayId:
          type: string
          example: vcp-gateway
        traceId:
          type: string
          example: agwop_1234567890abcdef123456
        operationName:
          type: string
          example: tool.invoke
        retryAfterMs:
          type: integer
        toolStatus:
          type: string
          enum:
            - completed
            - accepted
            - waiting_approval
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - code
        - meta
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
        code:
          type: string
          example: AGW_FORBIDDEN
        details:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    RequestContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
          example: native
    AuthContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
        adapter:
          type: string
        gatewayId:
          type: string
        authMode:
          type: string
        authSource:
          type: string
        roles:
          type: array
          items:
            type: string
...
    ToolInvokeRequest:
      type: object
      required:
        - args
        - requestContext
      properties:
        args:
          type: object
          additionalProperties: true
        options:
          type: object
          properties:
            idempotencyKey:
              type: string
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
    JobObject:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum:
            - accepted
            - running
            - waiting_approval
            - completed
            - failed
            - cancelled
        operation:
          type: string
        target:
          type: object
          nullable: true
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        authContext:
          $ref: "#/components/schemas/AuthContext"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        failedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        terminal:
          type: boolean
    ToolInvokeData:
      type: object
      properties:
        toolName:
          type: string
        result:
          type: object
          additionalProperties: true
        job:
          $ref: "#/components/schemas/JobObject"
        runtime:
          type: object
          additionalProperties: true
        audit:
          type: object
          additionalProperties: true
      example:
        toolName: ProtectedTool
        job:
          jobId: job_001
          status: waiting_approval
          operation: tool.invoke
          target:
            type: tool
            id: ProtectedTool
          metadata:
            toolName: ProtectedTool
          authContext:
            requestId: req-tool-001
            sessionId: sess-tool-001
            agentId: Ariadne
            runtime: native
            gatewayId: gw-prod
            authMode: gateway_key
            authSource: x-agent-gateway-key
          createdAt: "2026-04-19T12:00:00.000Z"
          updatedAt: "2026-04-19T12:00:00.000Z"
          terminal: false
        runtime:
          deferred: true
          status: waiting_approval
        audit:
          approvalUsed: true
          distributed: false
    ToolInvokeEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/ToolInvokeData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    JobData:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/JobObject"
    JobEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/JobData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    JobCancelRequest:
      type: object
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
...